feat: change in how FrameMaker implement classes are generated
Previous versions of Processing allowed to define multiple classes in a single .pde file, but this is no longer allowed in upgrades.

As the number of files increases, it becomes difficult to manage, so we added a class to generate the FrameMaker implementation class, and changed the generation of the FrameMaker implementation classes to instantiate an anonymous class using a method of the added class.

Other changes include:

- Made it easier to change the video file output path